requirements.txt
aiogram==3.20.0.post0
aiohttp>=3.8.0
aiosqlite>=0.17.0
python-dotenv>=1.0.0
asyncpg>=0.27.0
